externalize class
